+++ docs/changelog.md
@@ -12,6 +12,7 @@
 + Add "reduce" function for synthesis target path.
 + Ensure close the QThread before accept the dialog.
 + Remove the QMutex lock, which caused crash on Windows platform.
++ Improve the pattern of path file format.
 
 **Development**: